working. working with one threadpool, changing to one to server connection and threadpool for commands

+++ src/main/java/com/example/remotejoystick17/view_model/ViewModel.java
@@ -27,7 +27,7 @@ public class ViewModel extends Observable implements Observer {
         model.elevator(e);
     }
 
-    public void setRudder(float r) {
+    public void setRudder(double r) {
         model.rudder(r);
     }
